Fix issue with internal builder
When building internally, we split the build procedure in two phases,
the java part and the native part. In the native part we employ the "mx
build --only..." option which builds only the given project without
trying to build its dependencies as well. Unfortunately with 21.2 this
appears to cause the following issue:

With the new "Module support" mx automatically detects and adds the
"-add-exports" that we need, but when using "mx build --only" mx ignores
any dependencies and thus it doesn't add any "-add-exports"...

This patch explicitly exports any modules we need.