Improve release script performance by filtering using milestone
Filtering PRs by checking if they are merged and then by milestone is
really slow, since for each PR github-api synchronizes with github to
see if the PR is merged (the info is not cached).

Milestone info on the other hand are already available in the
`GHPullRequest` objects, thus filtering PRs with milestone is almost
instant, then we only pay the overhead of checking if each PR is
merged only for a few PRs, usually less than 10 per release.