Better handling of MANIFEST.MF attributes
Previously, the build script would add these two attributes irrespective
whether or not they'd exist in the Graal artefacts:

  Specification-Version
  Implementation-Version

With this patch, only updates to the Implementation-Version attribute
are done based on the maven version in use. Specification-Version is
only added if it's not yet there in a jar or source jar file.

Closes: https://github.com/Karm/mandrel-integration-tests/issues/270